pyproject.toml
[project]
name = "math-tagger-mlops"
version = "1.1.0"
description = "Multilabel classification of mathematics problems using PyTorch and MLOps tools"
readme = "README.md"
authors = [
    { name = "Ilya E", email = "elushoviv@gmail.com" }
]
requires-python = ">=3.11"
dependencies = [
    "catboost>=1.2.8",
    "datasets>=4.4.2",
    "dvc>=3.65.0",
    "hydra-core>=1.3.2",
    "joblib>=1.5.3",
    "lightning>=2.6.0",
    "mlflow>=3.8.1",
    "numpy>=2.4.0",
    "onnx>=1.20.0",
    "onnxruntime>=1.23.2",
    "onnxscript>=0.5.7",
    "pandas>=2.3.3",
    "scikit-learn>=1.8.0",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
]

[tool.hatch.build.targets.wheel]
packages = ["math_classifier"]

[project.scripts]
math-tagger-mlops = "math_classifier.commands:main"


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "black>=25.12.0",
    "flake8>=7.3.0",
    "isort>=7.0.0",
    "pre-commit>=4.5.1",
    "prettier>=0.0.7",
]
